requirements.txt
langchain>=0.3.0
langchain-ollama>=0.3.0
langgraph>=0.4.0
chainlit>=2.0.0
tavily-python>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
markdown>=3.5.0
